Board notes from today's meeting.
These notes contain two of the three classes we need to make.
TaskTable.java will represent a collection of the tasks the user has
created. ClassTable.java will keep track of the student's courses. We
need to make a scheduler class that will handle additions to the
calendar and run the prioritizing algorithm.